# Conflicts:
#	out/production/Lemonade Stand/com/imerir/Lemonade.class
#	out/production/Lemonade Stand/com/imerir/Money.class
#	src/com/imerir/Lemonade.java
#	src/com/imerir/Main.java
#	src/com/imerir/Money.java

+++ src/com/imerir/Lemonade.java
@@ -4,23 +4,27 @@ public class Lemonade {
 
     private int numLemonade= 0;
     private int numAds = 0;
-    private int costLemonage = 0;
+    private double costLemonage = 0;
 
-    public Lemonade(int numLemonade, int numAds, int costLemonage) {
+    /* fonction principale permetant de crée les lemonade  */
+    public Lemonade(int numLemonade, int numAds, double costLemonage) {
         this.numLemonade = numLemonade;
         this.numAds = numAds;
         this.costLemonage= costLemonage;
     }
 
+    /* fonction permetant de retournée le nombre de lemonade a vendre */
     public int getNumLemonade() {
         return numLemonade;
     }
 
+    /* fonction permetant de retournée le nombre d'annonce a fair */
     public int getNumAds() {
         return numAds;
     }
 
-    public int getCostLemonage() {
+    /* fonction permetant de retournée le prix d'une lemonade */
+    public double getCostLemonage() {
         return costLemonage;
     }
 }
